refactor(hooks): extract auto-select default category into shared hook
DRY up identical useEffect in Category.tsx and FloatingNavigatorContainer.tsx
into useAutoSelectDefaultCategory. Fix falsy category ID check in
FloatingNavigator dropdown and remove dead null fallback in optimistic update.
